#13: Provide a general Client init so Clients forget something. / Move there also the ClientWebsocket connection.
